Move structs to header
